+++ src/Pages/Community_Main/CommunityMain.scss
@@ -51,7 +51,7 @@
     right: 0;
     left: 0;
     width: 100%;
-    height: 350%;
+    z-index: 1000;
 
     &.active {
       opacity: 0.3;
