+++ src/Pages/Community_Main/FeedUnit.js
@@ -1,54 +1,22 @@
 import React, { Component } from 'react';
 import './FeedUnit.scss';
+import itemInfos from './itemInfos';
+import profileImgs from './profileImgs';
 
 export default class FeedUnit extends Component {
   state = {
     itemNum: Math.floor(Math.random() * 3 + 1),
+    imgNum: Math.floor(Math.random() * 3 + 1),
     newItemArray: [],
   };
 
   componentDidMount() {
-    const itemInfos = [
-      {
-        product: 'shoes',
-        brandName: ['르꼬끄', '아디다스', '락포트', '컨버스'],
-        iconUrl:
-          'https://www.flaticon.com/svg/vstatic/svg/892/892453.svg?token=exp=1614231066~hmac=7d7379daf436ed06a40c0a42575404f2',
-      },
-      {
-        product: 'outer',
-        brandName: ['리버시블', '메종미네드', '23.65', '널디'],
-        iconUrl:
-          'https://www.flaticon.com/svg/vstatic/svg/1985/1985620.svg?token=exp=1614231192~hmac=0dd7ba37e981dc56cf117290f9a01140',
-      },
-      {
-        product: 'pants',
-        brandName: ['에이들', '캘빈클라인', '무아무아', '소나'],
-        iconUrl:
-          'https://www.flaticon.com/svg/vstatic/svg/884/884485.svg?token=exp=1614231229~hmac=107317e74fceb865309aab36f41efc6a',
-      },
-      {
-        product: 'Top',
-        brandName: ['텐블레이드', '마크엠', '커버낫', '청남방'],
-        iconUrl:
-          'https://www.flaticon.com/svg/vstatic/svg/3637/3637365.svg?token=exp=1614231369~hmac=1d8f6b69994aa8f1bc81d4092d2f91a9',
-      },
-      {
-        product: 'watch',
-        brandName: ['논논(NoneNone)', '롤렉스', '와치'],
-        iconUrl:
-          'https://www.flaticon.com/svg/vstatic/svg/878/878913.svg?token=exp=1614231333~hmac=9701e3e495665b6b6889e393e768c881',
-      },
-    ];
-
     this.setState({
       newItemArray: itemInfos.splice(0, this.state.itemNum),
     });
   }
 
   render() {
-    console.log(this.state.newItemArray);
-
     const {
       key,
       username,
@@ -60,6 +28,8 @@ export default class FeedUnit extends Component {
       commentsNum,
       createdTime,
     } = this.props;
+
+    console.log(profileImgs[this.state.imgNum]);
 
     return (
       <div className="FeedUnit" key={key}>
@@ -99,10 +69,7 @@ export default class FeedUnit extends Component {
 
         <section className="summary">
           <div className="summaryFeed">
-            <img
-              src="https://usercontents-c.styleshare.io/images/48380668/120x120"
-              alt="profilepicture"
-            />
+            <img src={profileImgs[this.state.imgNum]} alt="profilepicture" />
             <div className="InnerBox">
               <div className="topText">
                 <span>{username}</span>
@@ -134,10 +101,7 @@ export default class FeedUnit extends Component {
           if (comment.user) {
             return (
               <div className="comments">
-                <img
-                  src="https://usercontents-c.styleshare.io/images/48380668/120x120"
-                  alt="profile"
-                />
+                <img src={profileImgs[this.state.itemNum]} alt="profile" />
                 <div>
                   <p>
                     <span>{comment.user}</span> {comment.content}
