+++ src/Pages/Community_Main/profileImgs.js
@@ -1,0 +1,10 @@
+const profileImgs = [
+  'https://ca.slack-edge.com/TH0U6FBTN-U01G4E9JEE5-8309fc852e7c-512',
+  'https://ca.slack-edge.com/TH0U6FBTN-U01FNNKKDEH-f7089670b696-512',
+  'https://ca.slack-edge.com/TH0U6FBTN-U01H0EEP93Q-ff9b01a3cff7-512',
+  'https://ca.slack-edge.com/TH0U6FBTN-U01FY0JNGS0-15b25f7d0e07-512',
+  'https://ca.slack-edge.com/TH0U6FBTN-U01G4HFRH0R-04274015e67d-512',
+  'https://ca.slack-edge.com/TH0U6FBTN-U01FRR5PC9Y-f7a3531bab82-512',
+];
+
+export default profileImgs;
